requirements.txt
bitarray==3.4.2
bitstring==4.3.1
cffi==1.17.1
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
fonttools==4.58.4
imageio==2.37.0
kiwisolver==1.4.8
lazy_loader==0.4
matplotlib==3.10.3
networkx==3.5
numpy==2.3.1
opencv-python==4.11.0.86
osxmetadata==1.4.0
packaging==25.0
pillow==11.2.1
py-applescript==1.0.3
pycparser==2.22
pymediainfo==7.0.1
pyobjc-core==11.1
pyobjc-framework-AppleScriptKit==11.1
pyobjc-framework-AppleScriptObjC==11.1
pyobjc-framework-AVFoundation==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-CoreAudio==11.1
pyobjc-framework-CoreMedia==11.1
pyobjc-framework-CoreServices==11.1
pyobjc-framework-FSEvents==11.1
pyobjc-framework-Quartz==11.1
pyobjc-framework-ScriptingBridge==11.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scikit-image==0.25.2
scipy==1.16.0
six==1.17.0
tifffile==2025.6.11
tqdm==4.67.1
xattr==1.1.4
